+++ src/app/psat/explore-opportunities/explore-opportunities-form/system-data-form/system-data-form.component.html
@@ -1,0 +1,116 @@
+<div class="checkbox">
+    <label for="general">
+        <input type="checkbox" id="general" value="true" (focus)="focusField('systemData')" (change)="toggleSystemData()" [(ngModel)]="showSystemData"
+        /> Adjust System Data
+    </label>
+</div>
+<ul *ngIf="showSystemData">
+    <li>
+        <div class="checkbox">
+            <label for="modifyCost">
+                <input type="checkbox" id="modifyCost" value="true" (focus)="focusField('modifyCost')" (change)="toggleCost()"
+                    [(ngModel)]="showCost" /> Modify Cost
+            </label>
+        </div>
+        <form *ngIf="showCost">
+            <div class="form-group">
+                <label for="baselineCost">Baseline Cost</label>
+                <div class="input-group">
+                    <input name="baselineCost" type="number" step="any" id="baselineCost" class="form-control" (focus)="focusField('modifyCost')"
+                        (input)="checkCost(1)" [(ngModel)]="psat.inputs.cost_kw_hour" />
+                    <span class="input-group-addon units">$/kWH</span>
+                </div>
+                <span class="alert-danger pull-right small" *ngIf="costError1 !== null">{{costError1}}</span>
+            </div>
+            <div class="form-group">
+                <label for="modificationCost">Modification Cost</label>
+                <div class="input-group">
+                    <input name="modificationCost" type="number" step="any" id="modificationCost" class="form-control" (focus)="focusField('modifyCost')"
+                        [(ngModel)]="psat.modifications[exploreModIndex].psat.inputs.cost_kw_hour" (input)="checkCost(2)" />
+                    <span class="input-group-addon units">($/kWh)</span>
+                </div>
+                <span class="alert-danger pull-right small" *ngIf="costError2 !== null">{{costError2}}</span>
+            </div>
+        </form>
+    </li>
+    <li>
+        <div class="checkbox">
+            <label for="modifyFlowRate">
+                <input type="checkbox" id="modifyFlowRate" value="true" (focus)="focusField('flowRate')" (change)="toggleFlowRate()"
+                    [(ngModel)]="showFlowRate" /> Modify Flow Rate
+            </label>
+        </div>
+        <form *ngIf="showFlowRate">
+            <div class="form-group">
+                <label for="baselineFlowRate">Baseline Flow Rate</label>
+                <div class="input-group">
+                    <input name="baselineFlowRate" type="number" step="any" id="baselineFlowRate" class="form-control" (focus)="focusField('flowRate')"
+                        (input)="checkFlowRate(1)" [(ngModel)]="psat.inputs.flow_rate" />
+                    <span class="input-group-addon units" [innerHTML]="getUnit(settings.flowMeasurement)"></span>
+                    <!--<span class="input-group-addon units">{{settings.flowMeasurement}}</span>-->
+                </div>
+                <span class="alert-danger pull-right small" *ngIf="flowRateError1 !== null">{{flowRateError1}}</span>
+            </div>
+            <div class="form-group">
+                <label for="modificationFlowRate">Modification Flow Rate</label>
+                <div class="input-group">
+                    <input name="modificationFlowRate" type="number" step="any" id="modificationFlowRate" class="form-control" (focus)="focusField('flowRate')"
+                        [(ngModel)]="psat.modifications[exploreModIndex].psat.inputs.flow_rate" (input)="checkFlowRate(2)" />
+                    <span class="input-group-addon units" [innerHTML]="getUnit(settings.flowMeasurement)"></span>
+                    <!--<span class="input-group-addon units">{{settings.flowMeasurement}}</span>-->
+                </div>
+                <span class="alert-danger pull-right small" *ngIf="flowRateError2 !== null">{{flowRateError2}}</span>
+            </div>
+        </form>
+    </li>
+    <li>
+        <div class="checkbox">
+            <label for="modifyHead">
+                <input type="checkbox" id="modifyHead" value="true" (focus)="focusField('head')" (change)="toggleHead()" [(ngModel)]="showHead"
+                /> Modify Head
+            </label>
+        </div>
+        <form *ngIf="showHead">
+            <div class="form-group">
+                <label for="baselineHead">Baseline head</label>
+                <div class="input-group">
+                    <input name="baselineHead" type="number" step="any" id="baselineHead" class="form-control" (focus)="focusField('head')" (input)="calculate()"
+                        [(ngModel)]="psat.inputs.head" />
+                    <span class="input-group-addon units" [innerHTML]="getUnit(settings.distanceMeasurement)"></span>
+
+                </div>
+            </div>
+            <div class="form-group">
+                <label for="modificationHead">Modification head</label>
+                <div class="input-group">
+                    <input name="modificationHead" type="number" step="any" id="modificationHead" class="form-control" (focus)="focusField('head')"
+                        (input)="calculate()" [(ngModel)]="psat.modifications[exploreModIndex].psat.inputs.head" />
+                    <span class="input-group-addon units" [innerHTML]="getUnit(settings.distanceMeasurement)"></span>
+                </div>
+            </div>
+        </form>
+    </li>
+    <li>
+        <div class="checkbox">
+            <label for="modifyOperatingFraction">
+                <input type="checkbox" id="modifyOperatingFraction" value="true" (focus)="focusField('operatingFraction')" (change)="toggleOperatingFraction()"
+                    [(ngModel)]="showOperatingFraction" /> Modify Operating Fraction
+            </label>
+        </div>
+        <form *ngIf="showOperatingFraction">
+            <div class="form-group">
+                <label for="baselineOperatingFraction">Baseline operating fractiong</label>
+                <input name="baselineOperatingFraction" type="number" step="any" id="baselineOperatingFraction" class="form-control" (focus)="focusField('operatingFraction')"
+                    (input)="checkOpFraction(1)" [(ngModel)]="psat.inputs.operating_fraction" />
+                <span class="alert-danger pull-right small" *ngIf="opFractionError1 !== null">{{opFractionError1}}</span>
+            </div>
+            <div class="form-group">
+                <label for="modificationOperatingFraction">Modification operating fraction</label>
+                <input name="modificationOperatingFraction" type="number" step="any" id="modificationOperatingFraction" class="form-control"
+                    (focus)="focusField('operatingFraction')" (input)="checkOpFraction(2)" [(ngModel)]="psat.modifications[exploreModIndex].psat.inputs.operating_fraction"
+                />
+                <span class="alert-danger pull-right small" *ngIf="opFractionError2 !== null">{{opFractionError2}}</span>
+            </div>
+        </form>
+    </li>
+</ul>
