+++ src/app/psat/explore-opportunities/explore-opportunities.module.ts
@@ -7,6 +7,9 @@ import { ExploreOpportunitiesResultsComponent } from './explore-opportunities-re
 import { ExploreOpportunitiesSankeyComponent } from './explore-opportunities-sankey/explore-opportunities-sankey.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { SharedModule } from '../../shared/shared.module';
+import { SystemDataFormComponent } from './explore-opportunities-form/system-data-form/system-data-form.component';
+import { RatedMotorFormComponent } from './explore-opportunities-form/rated-motor-form/rated-motor-form.component';
+import { PumpDataFormComponent } from './explore-opportunities-form/pump-data-form/pump-data-form.component';
 
 @NgModule({
   imports: [
@@ -15,7 +18,16 @@ import { SharedModule } from '../../shared/shared.module';
     SharedModule,
     ReactiveFormsModule
   ],
-  declarations: [ExploreOpportunitiesComponent, ExploreOpportunitiesFormComponent, ExploreOpportunitiesHelpComponent, ExploreOpportunitiesResultsComponent, ExploreOpportunitiesSankeyComponent],
+  declarations: [
+    ExploreOpportunitiesComponent, 
+    ExploreOpportunitiesFormComponent, 
+    ExploreOpportunitiesHelpComponent, 
+    ExploreOpportunitiesResultsComponent, 
+    ExploreOpportunitiesSankeyComponent,
+    SystemDataFormComponent,
+    RatedMotorFormComponent,
+    PumpDataFormComponent
+  ],
   exports: [ExploreOpportunitiesComponent, ExploreOpportunitiesResultsComponent]
 })
 export class ExploreOpportunitiesModule { }
